# collections/SwaggerCollections.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Bytes4All",
        "description": "RestFul API of the Bytes4All application. The API allows users to upload and download e-books",
        "termsOfService": "https://edielsonassis.com.br/termos-a-serem-definidos",
        "contact": {
            "name": "Edielson Assis",
            "email": "grizos.ed@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/edielson-assis/bytes4all-api/blob/main/LICENSE"
        },
        "version": "v1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Finds all Books",
                "description": "Finds all Books",
                "operationId": "findAllBooks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            },
                            "application/x-yaml": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Authentication problem"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Updates a Book",
                "description": "Updates a Book by passing in a JSON, XML or YML representation of the book!",
                "operationId": "updateBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookUpdateRequest"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/BookUpdateRequest"
                            }
                        },
                        "application/x-yaml": {
                            "schema": {
                                "$ref": "#/components/schemas/BookUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Updated book",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookUpdateResponse"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookUpdateResponse"
                                }
                            },
                            "application/x-yaml": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookUpdateResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Something is wrong with the request"
                    },
                    "403": {
                        "description": "Forbidden - Authentication problem"
                    },
                    "404": {
                        "description": "Not found - Book not found"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                },
                "security": [
                    {
                        "bearer-key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Adds a new Book",
                "description": "Adds a new Book by passing in a JSON, XML or YML representation of the book!",
                "operationId": "createBook",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created book",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Something is wrong with the request"
                    },
                    "403": {
                        "description": "Forbidden - Authentication problem"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                },
                "security": [
                    {
                        "bearer-key": []
                    }
                ]
            }
        },
...
        "/api/v1/books/get/{name}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Finds a Book",
                "description": "Search for one or more books by name",
                "operationId": "findBookByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            },
                            "application/x-yaml": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Authentication problem"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                }
            }
        },
        "/api/v1/books/download/{filename}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "operationId": "downloadFile",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/books/{id}": {
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Deletes a Book",
                "description": "Deletes a Book by their Id",
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The Id of the book to find.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted book"
                    },
                    "403": {
                        "description": "Forbidden - Authentication problem"
                    },
                    "404": {
                        "description": "Not found - Book not found"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                },
                "security": [
                    {
                        "bearer-key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BookUpdateResponse": {
                "type": "object",
                "properties": {
                    "bookId": {
                        "type": "integer",
                        "description": "Unique identifier of the book",
                        "format": "int64",
                        "example": 1
                    },
                    "author": {
                        "type": "string",
                        "description": "Name of the author",
                        "example": "Robert C. Martin"
                    },
                    "launchDate": {
                        "type": "string",
                        "description": "The date the book was launched",
                        "format": "date",
                        "example": "2008-08-01"
                    },
                    "title": {
                        "type": "string",
                        "description": "Title of the book",
                        "example": "Clean Code: A Handbook of Agile Software Craftsmanship"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description of the book",
                        "example": "A book that teaches software developers how to write clean, maintainable, and efficient code."
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                },
                "description": "Represents a book"
            },
            "Links": {
                "type": "object",
                "additionalProperties": {
                    "$ref": "#/components/schemas/Link"
                }
            },
            "BookUpdateRequest": {
                "required": [
                    "author",
                    "bookId",
                    "description",
                    "launchDate",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "bookId": {
                        "type": "integer",
                        "description": "Unique identifier of the book",
                        "format": "int64",
                        "example": 1
                    },
                    "author": {
                        "pattern": "^[a-zA-ZÀ-ÿ\\s'.]+$",
                        "type": "string",
                        "description": "Name of the author",
                        "example": "Robert C. Martin"
                    },
                    "launchDate": {
                        "type": "string",
                        "description": "The date the book was launched",
                        "format": "date",
                        "example": "2008-08-01"
                    },
                    "title": {
                        "type": "string",
                        "description": "Title of the book",
                        "example": "Clean Code: A Handbook of Agile Software Craftsmanship"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description of the book",
                        "example": "A book that teaches software developers how to write clean, maintainable, and efficient code."
                    }
                },
                "description": "Represents a book"
            },
            "BookResponse": {
                "type": "object",
                "properties": {
                    "bookId": {
                        "type": "integer",
                        "description": "Unique identifier of the book",
                        "format": "int64",
                        "example": 1
                    },
                    "author": {
                        "type": "string",
                        "description": "Name of the author",
                        "example": "Robert C. Martin"
                    },
                    "launchDate": {
                        "type": "string",
                        "description": "The date the book was launched",
                        "format": "date",
                        "example": "2008-08-01"
                    },
                    "title": {
                        "type": "string",
                        "description": "Title of the book",
                        "example": "Clean Code: A Handbook of Agile Software Craftsmanship"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description of the book",
                        "example": "A book that teaches software developers how to write clean, maintainable, and efficient code."
                    },
                    "downloadUrl": {
                        "type": "string",
                        "description": "PDF files"
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    }
                },
                "description": "Represents a book"
            },
            "BookRequest": {
                "required": [
                    "author",
                    "description",
                    "file",
                    "launchDate",
                    "title"
                ],
                "type": "object",
                "properties": {
                    "author": {
                        "pattern": "^[a-zA-ZÀ-ÿ\\s'.]+$",
                        "type": "string",
                        "description": "Name of the author",
                        "example": "Robert C. Martin"
                    },
                    "launchDate": {
                        "type": "string",
                        "description": "The date the book was launched",
                        "format": "date",
                        "example": "2008-08-01"
                    },
                    "title": {
                        "type": "string",
                        "description": "Title of the book",
                        "example": "Clean Code: A Handbook of Agile Software Craftsmanship"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description of the book",
                        "example": "A book that teaches software developers how to write clean, maintainable, and efficient code."
                    },
                    "file": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "description": "Represents a book"
            },
...
            "Link": {
                "type": "object",
                "properties": {
                    "href": {
                        "type": "string"
                    },
                    "hreflang": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "deprecation": {
                        "type": "string"
                    },
                    "profile": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "templated": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer-key": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}